requirements.txt
Flask
pyTelegramBotAPI
PyGithub
toml
pendulum